agentic-ai/retail-demo/logistics_agent/pyproject.toml
[project]
name = "logistics_agent"
version = "1.0.0"
description = "Logiistics Agent"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk>=0.3.0",
    "starlette>=0.46.1",
    "uvicorn>=0.34.0",
    "click>=8.1.8",
    "google-adk>=1.7.0",
    "python-dotenv>=1.1.0",
    "asyncpg>=0.30.0",
    "litellm>=1.40.0",
    "toolbox_core",
    "googleapis-common-protos==1.71.0",
    "google.api",
    "a2a-sdk[http-server]",
    "google-api-core",
    "sqlalchemy",
    "aiohttp",
]